refactor: Service, Dao 인터페이스 분리
